+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Support SIMILAR TO statement.
+
 ### Changed
 
 - "More" keyword is cleared on every page request.